requirements.txt
python-dotenv
opencv-python
PyQt6
imutils
tensorflow==2.13.*
tensorflow-addons
tf_slim
matplotlib
numpy
tensorflow_hub
#tensorflow-object-detection-api
